move message parsing functions to @cumulus/common
