+++ packages/common/message.js
@@ -3,6 +3,8 @@ const get = require('lodash.get');
 const merge = require('lodash.merge');
 const uuidv4 = require('uuid/v4');
 
+const { isNil } = require('./util');
+
 const {
   getS3Object,
   parseS3Uri
@@ -62,6 +64,52 @@ const buildMeta = ({
 const getQueueNameByUrl = (message, queueUrl) => {
   const queues = get(message, 'meta.queues', {});
   return findKey(queues, (value) => value === queueUrl);
+};
+
+
+/**
+ * Get the queue name from a workflow message.
+ *
+ * @param {Object} message - A workflow message object
+ * @returns {string} - A queue name
+ */
+const getQueueName = (message) => {
+  const queueName = get(message, 'cumulus_meta.queueName');
+  if (isNil(queueName)) {
+    throw new Error('cumulus_meta.queueName not set in message');
+  }
+  return queueName;
+};
+
+/**
+ * Get the maximum executions for a queue.
+ *
+ * @param {Object} message - A workflow message object
+ * @param {string} queueName - A queue name
+ * @returns {number} - Count of the maximum executions for the queue
+ */
+const getMaximumExecutions = (message, queueName) => {
+  const maxExecutions = get(message, `meta.queueExecutionLimits.${queueName}`);
+  if (isNil(maxExecutions)) {
+    throw new Error(`Could not determine maximum executions for queue ${queueName}`);
+  }
+  return maxExecutions;
+};
+
+/**
+ * Determine if there is a queue and queue execution limit in the message.
+ *
+ * @param {Object} message - A workflow message object
+ * @returns {boolean} - True if there is a queue and execution limit.
+ */
+const hasQueueAndExecutionLimit = (message) => {
+  try {
+    const queueName = getQueueName(message);
+    getMaximumExecutions(message, queueName);
+  } catch (err) {
+    return false;
+  }
+  return true;
 };
 
 /**
@@ -124,6 +172,9 @@ function buildQueueMessageFromTemplate({
 module.exports = {
   buildCumulusMeta,
   buildQueueMessageFromTemplate,
+  getQueueNameByUrl,
+  getQueueName,
+  getMaximumExecutions,
   getMessageFromTemplate,
-  getQueueNameByUrl
+  hasQueueAndExecutionLimit
 };
