+++ packages/common/package.json
@@ -58,6 +58,7 @@
     "lodash.curry": "^4.1.1",
     "lodash.findkey": "^4.6.0",
     "lodash.flowright": "^3.5.0",
+    "lodash.get": "^4.4.2",
     "lodash.isfunction": "^3.0.9",
     "lodash.isnil": "^4.0.0",
     "lodash.isnumber": "^3.0.3",
